+++ src/lib/components/modal/contract/EditContractDetails.tsx
@@ -1,0 +1,21 @@
+import type { ContractInfo } from "lib/stores/contract";
+
+import { ContractDetailsTemplate } from "./ContractDetailsTemplate";
+
+interface EditContractDetailsProps {
+  contractInfo: ContractInfo;
+  triggerElement: JSX.Element;
+}
+export const EditContractDetails = ({
+  contractInfo,
+  triggerElement,
+}: EditContractDetailsProps) => {
+  return (
+    <ContractDetailsTemplate
+      title="Edit Contract Details"
+      subtitle="Filled information below will be saved on Celatone only and able to edit later."
+      contractInfo={contractInfo}
+      triggerElement={triggerElement}
+    />
+  );
+};